[DeviceMesh] Rename _device_mesh.py to device_mesh.py to prepare for beta (#115193)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/115193

Rename _device_mesh.py to device_mesh.py, update all callsites, add documentation.
We created stubs for public class and methods in torch.distributed.device_mesh so that torch.distributed.device_mesh can be imported with or without distributed is available().

Original diff reverted: D51629761
Original PR reverted: https://github.com/pytorch/pytorch/pull/115099
Prior to landing, CI signals are all passed. Shipit added the "ci/trunk" label to the PR and DID NOT wait for it and went ahead committing. More context can be found in the reverted PR above.

Test Plan: CI.

Differential Revision: D51861018

fbshipit-source-id: dc7b26cea7340d55498730123e82a42cef46ff55
