+++ torch/distributed/_tensor/device_mesh.py
@@ -1,4 +1,4 @@
-from torch.distributed._device_mesh import (  # noqa: F401
+from torch.distributed.device_mesh import (  # noqa: F401
     _get_device_handle,
     _mesh_resources,
     DeviceMesh,
