+++ torch/distributed/_tensor/ops/basic_strategy.py
@@ -3,7 +3,6 @@ from dataclasses import dataclass
 
 from typing import List, Tuple
 
-from torch.distributed._tensor.device_mesh import DeviceMesh
 from torch.distributed._tensor.op_schema import OpStrategy, PlacementStrategy
 from torch.distributed._tensor.placement_types import (
     _Partial,
@@ -12,6 +11,8 @@ from torch.distributed._tensor.placement_types import (
     Replicate,
     Shard,
 )
+
+from torch.distributed.device_mesh import DeviceMesh
 
 
 @dataclass
