+++ torch/distributed/_tensor/ops/matrix_ops.py
@@ -1,8 +1,6 @@
 # Copyright (c) Meta Platforms, Inc. and affiliates
 # implement matrix related ops for distributed tensor
 import torch
-
-from torch.distributed._tensor.device_mesh import DeviceMesh
 from torch.distributed._tensor.op_schema import OpSchema, OpStrategy, OutputSharding
 from torch.distributed._tensor.ops.basic_strategy import gen_einsum_strategies
 from torch.distributed._tensor.ops.common_rules import einop_rule
@@ -15,6 +13,8 @@ from torch.distributed._tensor.ops.utils import (
     register_prop_rule,
 )
 from torch.distributed._tensor.placement_types import DTensorSpec
+
+from torch.distributed.device_mesh import DeviceMesh
 
 aten = torch.ops.aten
